Added Code to Multiple Files for better Readability
